fix: aggregator restart missing events (#737)
* chore: deal with rustc 1.89 mismatched_lifetime_syntaxes warnings

* fix: aggregator idempotency

restart from the beginning of conclusion events to avoid skipping on restarts. need to figure out why memory table contains conclusion_event_order that is less than the data written to event_states still, and hopefully restart where we left off. for now, we just ignore things we've seen before.

* fix: ensure conclusion event batch is written flushed from cache as a whole

we split the batch into models/mids and were writing both pieces to disk separately. The entire batch is orded by conclusion_event_order, but each group could overlap. So we could flush the cache for the first write and then we'd have a cache that was "going backward" in conclusion_event_order. On restart, those events were missed and events failed to aggregate correctly

* test: fix comment and modify test to better replicate ordering bug

* chore(ci): add fmt/clippy

* chore(ci): use newer jq image

* chore(ci): use alpine image for jq

* chore(ci): install bash in init container

---------

Co-authored-by: Mohsin Zaidi <mohsinrzaidi@gmail.com>